Add state to newly created JSON file
